31 Fix: Inclusion using class naming deprecated
Reflect Moodle 4.5's changes to Filter API (MDL-82427).
